Refactored duplicate code from `train_dqn()` and `run_dqn()` into a separate `process_episode()` function.
Moved `DQN` and `DQNAgent` to a separate file to separate the neural network logic from the management logic.
